+++ Source/equipment/scutum-base.js
@@ -1,13 +1,13 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { addBlock, removeModifier } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Ice Ward", "Grant @{block} block to an ally and yourself", "Block x@{critBonus}", "Water", effect, ["Heavy Ice Ward", "Sweeping Ice Ward"])
-	.setCategory("Spell")
+module.exports = new EquipmentTemplate("Scutum", "Grant @{block} block to an ally and yourself", "Block x@{critBonus}", "Fire", effect, ["Heavy Scutum", "Sweeping Scutum"])
+	.setCategory("Armor")
 	.setTargetingTags({ target: "single", team: "delver" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
 	.setCost(200)
 	.setUses(10)
-	.setBlock(50);
+	.setBlock(75);
 
 function effect([target], user, isCrit, adventure) {
 	let { element, modifiers: [elementStagger], block, critBonus } = module.exports;